Don't read CS1 and CS2 values when asking for motor status (as response to wireless message), instead read it the loop and only send previously saved value when asked

+++ moving-sculptures/modules/motor.h
@@ -12,8 +12,8 @@ struct motorParameters {
 struct motorStatus {
 	int speed;
 	byte direction;
-	byte cs1;
-	byte cs2;
+	int cs1;
+	int cs2;
 	int battery;
 	bool disabled;
 };
